o  fixed checkout tasks and config service.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497933 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/main/java/org/apache/maven/continuum/configuration/DefaultConfigurationService.java
@@ -170,7 +170,7 @@ public class DefaultConfigurationService
     public File getBuildOutputFile( long buildId, long projectId )
         throws ConfigurationException
     {
-        File dir = new File( getBuildOutputDirectory(), Integer.toString( projectId ) );
+        File dir = new File( getBuildOutputDirectory(), Long.toString( projectId ) );
 
         try
         {
